requirements.txt
flask==3.0.0
flask-cors==4.0.0
pandas==2.1.4
psycopg2-binary==2.9.9
sqlalchemy==2.0.25
python-dotenv==1.0.0
werkzeug==3.0.1